+++ .github/zz_generated.windows-code-signing.sh
@@ -2,7 +2,7 @@
 
 # DO NOT EDIT. Generated with:
 #
-#    devctl@5.14.0
+#    devctl@6.5.0
 #
 
 APPLICATION=$1