+++ cmd/gitops/add/base/flag.go
@@ -11,16 +11,21 @@ import (
 )
 
 const (
-	flagProvider = "provider"
+	flagAzureSubscriptionID = "azure-subscription-id"
+	flagProvider            = "provider"
+	flagRegion              = "region"
 )
 
 type flag struct {
-	Provider string
+	Provider            string
+	Region              string
+	AzureSubscriptionID string
 }
 
 func supportedProviders() []string {
 	return []string{
 		key.ProviderCAPA,
+		key.ProviderCAPZ,
 		key.ProviderGCP,
 		key.ProviderOpenStack,
 	}
@@ -28,6 +33,10 @@ func supportedProviders() []string {
 
 func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.Provider, flagProvider, "", fmt.Sprintf("Installation infrastructure provider, supported values: %s", strings.Join(supportedProviders(), ", ")))
+	cmd.Flags().StringVar(&f.Region, flagRegion, "", "AWS/Azure/GCP region where cluster will be created")
+
+	// Azure only
+	cmd.Flags().StringVar(&f.AzureSubscriptionID, flagAzureSubscriptionID, "", "Azure subscription ID")
 }
 
 func (f *flag) Validate() error {