+++ cmd/template/cluster/error.go
@@ -21,3 +21,7 @@ var invalidFlagError = &microerror.Error{
 func IsInvalidFlag(err error) bool {
 	return microerror.Cause(err) == invalidFlagError
 }
+
+var templateFlagNotImplemented = &microerror.Error{
+	Kind: "templateFlagsNotImplementedError",
+}