+++ cmd/template/cluster/provider/capz.go
@@ -1,111 +1,200 @@
 package provider
 
 import (
-	"bytes"
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
 	"io"
 	"text/template"
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
+	"sigs.k8s.io/yaml"
 
-	"github.com/giantswarm/kubectl-gs/v2/cmd/template/cluster/provider/templates/azure"
+	k8smetadata "github.com/giantswarm/k8smetadata/pkg/label"
+
+	"github.com/giantswarm/kubectl-gs/v2/cmd/template/cluster/provider/templates/capz"
 	"github.com/giantswarm/kubectl-gs/v2/internal/key"
+	templateapp "github.com/giantswarm/kubectl-gs/v2/pkg/template/app"
 )
 
-func WriteCAPZTemplate(ctx context.Context, client k8sclient.Interface, out io.Writer, config ClusterConfig) error {
-	var err error
+const (
+	DefaultAppsAzureRepoName = "default-apps-azure"
+	ClusterAzureRepoName     = "cluster-azure"
+)
 
-	var sshSSOPublicKey string
-	{
-		sshSSOPublicKey, err = key.SSHSSOPublicKey(ctx, client.CtrlClient())
-		if err != nil {
-			return microerror.Mask(err)
-		}
-	}
-
-	var ignitionBase64 string
-	{
-		data := struct {
-			SystemdUnits []struct {
-				Name     string
-				Contents string
-			}
-		}{
-			SystemdUnits: []struct {
-				Name     string
-				Contents string
-			}{
-				{
-					Name:     "set-bastion-ready.timer",
-					Contents: jsonEscape(key.CapzSetBastionReadyTimer),
-				},
-				{
-					Name:     "set-bastion-ready.service",
-					Contents: jsonEscape(key.CapzSetBastionReadyService),
-				},
-			},
-		}
-
-		var tpl bytes.Buffer
-		t := template.Must(template.New(config.FileName).Parse(fmt.Sprintf(key.BastionIgnitionTemplate, config.Name, key.BastionSSHDConfigEncoded(), base64.StdEncoding.EncodeToString([]byte(sshSSOPublicKey)))))
-		err = t.Execute(&tpl, data)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		ignitionBase64 = base64.StdEncoding.EncodeToString(tpl.Bytes())
-	}
-
-	data := struct {
-		BastionIgnitionSecretBase64 string
-		BastionVMSize               string
-		Description                 string
-		KubernetesVersion           string
-		Name                        string
-		Namespace                   string
-		SSHDConfig                  string
-		SSOPublicKey                string
-		Organization                string
-		PodsCIDR                    string
-		Version                     string
-		VMSize                      string
-	}{
-		BastionIgnitionSecretBase64: ignitionBase64,
-		BastionVMSize:               "Standard_D2_v3",
-		Description:                 config.Description,
-		KubernetesVersion:           "v1.19.9",
-		Name:                        config.Name,
-		Namespace:                   key.OrganizationNamespaceFromName(config.Organization),
-		Organization:                config.Organization,
-		PodsCIDR:                    config.PodsCIDR,
-		SSHDConfig:                  key.NodeSSHDConfigEncoded(),
-		SSOPublicKey:                sshSSOPublicKey,
-		Version:                     config.ReleaseVersion,
-		VMSize:                      "Standard_D4s_v3",
-	}
-
-	var templates []templateConfig
-	for _, t := range azure.GetTemplates() {
-		templates = append(templates, templateConfig(t))
-	}
-
-	err = runMutation(ctx, client, data, templates, out)
+func WriteCAPZTemplate(ctx context.Context, client k8sclient.Interface, output io.Writer, config ClusterConfig) error {
+	err := templateClusterCAPZ(ctx, client, output, config)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	return nil
+	err = templateDefaultAppsAzure(ctx, client, output, config)
+	return microerror.Mask(err)
 }
 
-func jsonEscape(i string) string {
-	b, err := json.Marshal(i)
-	if err != nil {
-		panic(err)
+func templateClusterCAPZ(ctx context.Context, k8sClient k8sclient.Interface, output io.Writer, config ClusterConfig) error {
+	appName := config.Name
+	configMapName := userConfigMapName(appName)
+
+	var configMapYAML []byte
+	{
+		flagValues := BuildCapzClusterConfig(config)
+
+		configData, err := capz.GenerateClusterValues(flagValues)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		userConfigMap, err := templateapp.NewConfigMap(templateapp.UserConfig{
+			Name:      configMapName,
+			Namespace: organizationNamespace(config.Organization),
+			Data:      configData,
+		})
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		userConfigMap.Labels = map[string]string{}
+		userConfigMap.Labels[k8smetadata.Cluster] = config.Name
+
+		configMapYAML, err = yaml.Marshal(userConfigMap)
+		if err != nil {
+			return microerror.Mask(err)
+		}
 	}
-	// Trim the beginning and trailing " character
-	return string(b[1 : len(b)-1])
+
+	var appYAML []byte
+	{
+		appVersion := config.App.ClusterVersion
+		if appVersion == "" {
+			var err error
+			appVersion, err = getLatestVersion(ctx, k8sClient.CtrlClient(), ClusterAzureRepoName, config.App.ClusterCatalog)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+		}
+
+		clusterAppConfig := templateapp.Config{
+			AppName:                 config.Name,
+			Catalog:                 config.App.ClusterCatalog,
+			InCluster:               true,
+			Name:                    ClusterAzureRepoName,
+			Namespace:               organizationNamespace(config.Organization),
+			Version:                 appVersion,
+			UserConfigConfigMapName: configMapName,
+		}
+
+		var err error
+		appYAML, err = templateapp.NewAppCR(clusterAppConfig)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	t := template.Must(template.New("appCR").Parse(key.AppCRTemplate))
+
+	err := t.Execute(output, templateapp.AppCROutput{
+		AppCR:               string(appYAML),
+		UserConfigConfigMap: string(configMapYAML),
+	})
+	return microerror.Mask(err)
+}
+
+func BuildCapzClusterConfig(config ClusterConfig) capz.ClusterConfig {
+	return capz.ClusterConfig{
+		Metadata: &capz.Metadata{
+			Name:         config.Name,
+			Description:  config.Description,
+			Organization: config.Organization,
+		},
+		ProviderSpecific: &capz.ProviderSpecific{
+			Location:       config.Region,
+			SubscriptionID: config.Azure.SubscriptionID,
+		},
+		Connectivity: &capz.Connectivity{
+			Bastion: &capz.Bastion{
+				Enabled:      true,
+				InstanceType: config.BastionInstanceType,
+			},
+		},
+		ControlPlane: &capz.ControlPlane{
+			InstanceType: config.ControlPlaneInstanceType,
+			Replicas:     3,
+		},
+	}
+}
+
+func templateDefaultAppsAzure(ctx context.Context, k8sClient k8sclient.Interface, output io.Writer, config ClusterConfig) error {
+	appName := fmt.Sprintf("%s-default-apps", config.Name)
+	configMapName := userConfigMapName(appName)
+
+	var configMapYAML []byte
+	{
+		flagValues := capz.DefaultAppsConfig{
+			ClusterName:  config.Name,
+			Organization: config.Organization,
+		}
+
+		configData, err := capz.GenerateDefaultAppsValues(flagValues)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		userConfigMap, err := templateapp.NewConfigMap(templateapp.UserConfig{
+			Name:      configMapName,
+			Namespace: organizationNamespace(config.Organization),
+			Data:      configData,
+		})
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		userConfigMap.Labels = map[string]string{}
+		userConfigMap.Labels[k8smetadata.Cluster] = config.Name
+
+		configMapYAML, err = yaml.Marshal(userConfigMap)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	var appYAML []byte
+	{
+		appVersion := config.App.DefaultAppsVersion
+		if appVersion == "" {
+			var err error
+			appVersion, err = getLatestVersion(ctx, k8sClient.CtrlClient(), DefaultAppsAzureRepoName, config.App.DefaultAppsCatalog)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+		}
+
+		var err error
+		appYAML, err = templateapp.NewAppCR(templateapp.Config{
+			AppName:                 appName,
+			Cluster:                 config.Name,
+			Catalog:                 config.App.DefaultAppsCatalog,
+			DefaultingEnabled:       false,
+			InCluster:               true,
+			Name:                    DefaultAppsAzureRepoName,
+			Namespace:               organizationNamespace(config.Organization),
+			Version:                 appVersion,
+			UserConfigConfigMapName: configMapName,
+			UseClusterValuesConfig:  true,
+			ExtraLabels: map[string]string{
+				k8smetadata.ManagedBy: "cluster",
+			},
+		})
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	t := template.Must(template.New("appCR").Parse(key.AppCRTemplate))
+
+	err := t.Execute(output, templateapp.AppCROutput{
+		UserConfigConfigMap: string(configMapYAML),
+		AppCR:               string(appYAML),
+	})
+	return microerror.Mask(err)
 }